Add a retry_sleep attribute in lamp_model module
This should fix a bug in the check for the unexpectedly off condition
where the CSC would wait the max_lamp_on_delay before retrying
which would lead to a condition where the CSC would fault during
a retry because timer did not account for this.
As a fix, I added an attribute for waiting a shorter time to retry
and accounted for it in the check.